ai-service/requirements.txt
# AI Service Dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LangChain ecosystem
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-core>=0.1.10

# Vector stores and embeddings
chromadb>=0.4.22
tiktoken>=0.5.2

# Document processing
pypdf2>=3.0.0
youtube-transcript-api>=0.6.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Search tools
tavily-python>=0.3.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
aiofiles>=23.2.1
